BAI_Tool/Monitor/requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.1.2
Werkzeug==3.0.1
reportlab==4.0.7
pandas==2.1.4
plotly==5.18.0
gunicorn==21.2.0
